% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{A function that takes in the name of a country in the form of a string and
returns the a graph that displays the population from the years 1950 - 2020.}
\usage{
CountryPopulation(country_name)
}
\arguments{
\item{country_name}{A string of the country name.}
}
\value{
A graph that displays the country's population data from 1950 - 2020.
}
\description{
A function that takes in the name of a country in the form of a string and
returns the a graph that displays the population from the years 1950 - 2020.
}
