% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{A series of past World Cup attendance records.}
\format{
a data frame with  observations
\describe{
  \item{WorldCup}{Country and year of past World Cups.}
  \item{Matches}{Total matches played.}
  \item{Totalattendance}{Total number of attendance of games.}
  \item{Averageattendance}{Average attendance per match.}
}
}
\source{
\url{'https://en.wikipedia.org/wiki/FIFA_World_Cup'}
}
\usage{
World_Cup
}
\description{
A series of past World Cup attendance records.
}
\keyword{datasets}
