% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{A series of world population data from 1950 - 2020.}
\format{
a data frame with  observations
\describe{
  \item{Country Name}{Country that observations originate from.}
  \item{1950 - 2020}{Year of population observations.}
}
}
\source{
\url{United Nations, Population Division}
}
\usage{
WorldPopulation
}
\description{
A series of world population data from 1950 - 2020.
}
\keyword{datasets}
